beautiful: eager loading
This commit in record low number of lines implements eager loading:
when the current item gets close to the edge of currently loaded
items, it 'recenters' the current buffer around the current item.

In theory at least, and also with good working servers and good networks,
this should allow user to never wait.

Signed-off-by: Zeljko <zeljko@zwr.fi>